Put one entity

+++ run.py
@@ -3,6 +3,7 @@ from digeiz import db
 
 
 if __name__ == '__main__':
-    db.drop_all()
+    if app.config["RESET_DB"]:
+        db.drop_all()
     db.create_all()
     app.run(debug=app.config["DEBUG"])
